requirements-base.txt
# Base dependencies (platform-agnostic)
# Install this first, then install platform-specific PyTorch from:
# - requirements-cuda.txt (Linux with NVIDIA GPU)
# - requirements-mps.txt (macOS Apple Silicon)
# - requirements-cpu.txt (CPU-only)

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Auth
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ML/AI (without PyTorch - install separately)
numpy>=1.26.0
transformers>=4.36.0,<5.0
accelerate>=0.25.0
huggingface_hub>=0.23.0,<1.0

# Speech/Audio
faster-whisper>=1.0.0
pyannote-audio==3.3.2
funasr>=1.3.0
demucs>=4.0.0
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0

# Video
opencv-python>=4.8.0
av>=12.0.0
scenedetect>=0.6.2

# Image
pillow>=10.1.0

# PDF Generation
reportlab>=4.0.0

# Utils
tqdm>=4.66.1
httpx>=0.25.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Dev (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.7.0
